refactor mq-client
