+++ lib/instance.js
@@ -1,3 +1,4 @@
+const debug = require('debug')('nats-instance:');
 const NATS = require('nats');
 
 let nats = null;
@@ -9,6 +10,25 @@ module.exports = function getInstance(natsConfig = {}) {
 
   if (!nats) {
     nats = NATS.connect({ url: `nats://${url}:${port}`, json: jsonMode });
+
+    nats.on('error', (err) => {
+      // exit process
+      debug('nats connection errored', err);
+      process.exit(1);
+    });
+
+    nats.on('disconnect', () => {
+      // exit process
+      debug('nats connection disconnected');
+      process.exit(1);
+    });
+
+    nats.on('close', () => {
+      // exit process
+      debug('nats connection closed');
+      process.exit(1);
+    });
+
     return nats;
   }
   return nats;
